Third Push + Création de la couche Repository et Service

+++ src/main/java/tn/esprit/projetspring1/Entities/Chambre.java
@@ -18,7 +18,7 @@ import java.util.Set;
 public class Chambre implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idChambre;
+    private long idChambre;
     private int numeroChambre;
     @Enumerated (EnumType.STRING)
     private TypeChambre typeC;
