+++ src/main/java/tn/esprit/projetspring1/Entities/Etudiant.java
@@ -19,7 +19,7 @@ import java.util.Set;
 public class Etudiant implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idEtudiant ;
+    private long idEtudiant ;
     private String nomEt;
     private String prenomEt;
     private int cin;
